backend/requirements.txt
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
cryptography>=42.0.0
dnspython==2.8.0
docker>=7.0.0
email-validator==2.3.0
fastapi==0.133.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg==3.3.2
psycopg-binary==3.3.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT[crypto]>=2.8.0
python-dotenv==1.2.1
SQLAlchemy==2.0.46
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
